# frontend/openapi/wst-openapi.yaml
openapi: 3.1.0
info:
  title: Workshop Management & Student Practical Training (WST) API
  version: 1.0.0
  summary: Contract-first REST API for the WST academic MVP.
  description: |
    Source of truth for the PostgreSQL schema, Node.js backend, automated tests, React client and documentation.
    Requirement prefix: WST. Every operation carries `x-requirement-id` and `x-priority`.

    ## Conventions
    - Base path `/api/v1`. JSON bodies (multipart only for attachment upload). Bearer JWT. Operations with `security: []` are public.
    - Identifiers are UUIDs. Human-readable numbers (`jobNumber`, `poNumber`, `invoiceNumber`, `certificateNumber`) are server-generated, unique, immutable, and never localized or reordered by the backend.
    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD`.
    - Money is `{amount, currency}`: `amount` is a decimal string (never a float) and `currency` an ISO-4217 code. One system currency is configured in finance settings; documents in another currency are rejected with 422. Part quantities are whole numbers. Rates and metric values are decimal strings.
    - Optional properties without a value are omitted, never null. `readOnly` properties are server-controlled. Request schemas forbid unknown properties, so clients cannot submit ids, audit fields, stock balances, invoice totals, or permissions.
...

    ## Collections
...

    ## Authorization
...

    ## Concurrency and idempotency
...

    ## Localization
...

    ## Errors and rate limiting
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      operationId: listVehicles
      summary: List vehicles
      description: "Requires `vehicles.read`. Sortable: plate, make, createdAt (default plate)."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Query'
        - { name: customerId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: plate, in: query, schema: { type: string, maxLength: 20 } }
        - { name: vin, in: query, schema: { type: string, maxLength: 17 } }
        - { name: make, in: query, schema: { type: string, maxLength: 60 } }
        - { name: model, in: query, schema: { type: string, maxLength: 60 } }
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, ARCHIVED] } }
      responses:
        '200': { description: Vehicles, content: { application/json: { schema: { $ref: '#/components/schemas/VehiclePage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Vehicles]
      operationId: createVehicle
      summary: Register a vehicle for a customer
      description: "Requires `vehicles.write`. A customer may own many vehicles. Plate and VIN must be unique among active vehicles (409 DUPLICATE_RESOURCE)."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.write]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VehicleCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /vehicles/{vehicleId}:
    parameters:
      - $ref: '#/components/parameters/VehicleId'
    get:
      tags: [Vehicles]
      operationId: getVehicle
      summary: Get vehicle detail with next-service information
      description: "Requires `vehicles.read`. Includes lastCompletedJob and nextService (earliest open reminder due date or mileage)."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.read]
      responses:
        '200': { description: Vehicle, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Vehicles]
      operationId: updateVehicle
      summary: Update or archive a vehicle
      description: "Requires `vehicles.write`. Mileage may only increase (422). Archiving is rejected with 409 RESOURCE_IN_USE while a non-delivered job exists."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.write]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VehicleUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Vehicle' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /vehicles/{vehicleId}/service-history:
    parameters:
      - $ref: '#/components/parameters/VehicleId'
    get:
      tags: [Vehicles]
      operationId: listVehicleServiceHistory
      summary: Completed-job history of a vehicle
      description: "Requires `vehicles.read`. Lists DELIVERED jobs only. Sortable: deliveredAt (default -deliveredAt)."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': { description: History, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceHistoryPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /vehicles/{vehicleId}/reminders:
    parameters:
      - $ref: '#/components/parameters/VehicleId'
    get:
      tags: [Vehicles]
      operationId: listServiceReminders
      summary: List service reminders of a vehicle
      description: "Requires `vehicles.read`. Reminders are manual records; automated sending is Stretch. Sortable: dueDate, createdAt."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - { name: status, in: query, schema: { type: string, enum: [OPEN, DONE, CANCELLED] } }
      responses:
        '200': { description: Reminders, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceReminderPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Vehicles]
      operationId: createServiceReminder
      summary: Record a service reminder
      description: "Requires `vehicles.write`. At least one of dueDate or dueMileage is required."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.write]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceReminderCreateRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceReminder' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /service-reminders/{reminderId}:
    parameters:
      - $ref: '#/components/parameters/ReminderId'
    patch:
      tags: [Vehicles]
      operationId: updateServiceReminder
      summary: Edit, complete or cancel a reminder
      description: "Requires `vehicles.write`. DONE and CANCELLED are terminal (409 INVALID_STATE_TRANSITION on further status change)."
      x-requirement-id: [WST-FR-03]
      x-priority: Must
      x-permissions: [vehicles.write]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceReminderUpdateRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceReminder' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token issued by login or refresh. The token identifies the user only; roles, permissions and scopes are resolved server-side on each request."
  parameters:
    Page: { name: page, in: query, description: "1-based page number.", schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, description: "Items per page, maximum 100.", schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, description: "Comma-separated fields; prefix - for descending. Only fields named by the operation are accepted.", schema: { type: string, maxLength: 100, pattern: '^-?[A-Za-z][A-Za-z0-9]*(,-?[A-Za-z][A-Za-z0-9]*)*$' } }
...
    Query: { name: q, in: query, description: "Case-insensitive text search.", schema: { type: string, minLength: 1, maxLength: 100 } }
...
    VehicleId: { name: vehicleId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    ReminderId: { name: reminderId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
  responses:
    BadRequest:
      description: "Malformed request, unknown filter or sort field, or invalid query parameter."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: "Missing, invalid or expired credentials."
      headers: { WWW-Authenticate: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: "Authenticated but lacking the required permission. Rows outside the caller's scope are reported as 404 instead."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: "Resource does not exist or is outside the caller's row scope."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: "Invalid state transition, missing customer approval, insufficient stock, concurrency conflict, separation-of-duties violation or duplicate business operation. See Error.code."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ValidationError:
      description: "Semantic validation failed; details identify each offending field."
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time, description: "RFC 3339 UTC timestamp." }
    DateOnly: { type: string, format: date }
...
    RecordMeta:
      type: object
      required: [id, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: '#/components/schemas/Uuid', readOnly: true }
        createdAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        updatedAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
        createdBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
        updatedBy: { $ref: '#/components/schemas/Uuid', readOnly: true }
    PageInfo:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties: { page: { type: integer, minimum: 1 }, pageSize: { type: integer, minimum: 1, maximum: 100 }, totalItems: { type: integer, minimum: 0 }, totalPages: { type: integer, minimum: 0 } }
...
    ErrorCode:
      type: string
      enum: [BAD_REQUEST, UNAUTHENTICATED, INVALID_CREDENTIALS, TOKEN_EXPIRED, FORBIDDEN, NOT_FOUND, VALIDATION_FAILED, RATE_LIMITED, INTERNAL_ERROR, SERVICE_UNAVAILABLE, INVALID_STATE_TRANSITION, VERSION_CONFLICT, DUPLICATE_RESOURCE, IDEMPOTENCY_CONFLICT, RESOURCE_IN_USE, CUSTOMER_APPROVAL_REQUIRED, JOB_STAGE_NOT_ALLOWED, JOB_ASSIGNMENT_REQUIRED, CHECKLIST_INCOMPLETE, QUALITY_CHECK_REQUIRED, INVOICE_REQUIRED, INSUFFICIENT_STOCK, REVERSAL_EXCEEDS_ISSUED, RESERVATION_INVALID, SCHEDULE_CONFLICT, CONFLICT_NOT_OVERRIDABLE, SEPARATION_OF_DUTIES_VIOLATION, DUPLICATE_APPROVAL, RECEIPT_EXCEEDS_ORDERED, PAYMENT_AMOUNT_MISMATCH, ASSESSMENT_LOCKED, CERTIFICATE_NOT_ELIGIBLE, ATTACHMENT_INVALID, ATTACHMENT_NOT_LINKABLE, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, EXPORT_NOT_READY, EXPORT_EXPIRED]
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        field: { type: string, description: "JSON Pointer to the offending request field." }
        code: { type: string, description: "Machine-readable detail code, for example REQUIRED, TOO_LONG, MILEAGE_LOWER_THAN_RECORDED." }
        message: { type: string }
        params: { type: object, additionalProperties: { type: string }, description: "Named values for message templates, for example available and requested quantities." }
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, description: "Human-readable message in the negotiated language." }
        details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
        requestId: { type: string, description: "Correlation id also present in logs and audit events." }
...
    NextService:
      type: object
      properties: { dueDate: { $ref: '#/components/schemas/DateOnly' }, dueMileage: { type: integer, minimum: 0 }, reminderId: { $ref: '#/components/schemas/Uuid' } }
    ServiceHistoryItem:
      type: object
      required: [jobId, jobNumber, serviceType, complaint, mileageAtIntake, deliveredAt]
      properties: { jobId: { $ref: '#/components/schemas/Uuid' }, jobNumber: { $ref: '#/components/schemas/JobNumber' }, serviceType: { $ref: '#/components/schemas/ServiceType' }, complaint: { type: string }, mileageAtIntake: { type: integer, minimum: 0 }, deliveredAt: { $ref: '#/components/schemas/Timestamp' } }
    Vehicle:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [customerId, plate, vin, make, model, year, mileage, mileageUnit, status]
          properties:
            customerId: { $ref: '#/components/schemas/Uuid' }
            plate: { type: string, minLength: 2, maxLength: 20 }
            vin: { type: string, pattern: '^[A-HJ-NPR-Z0-9]{17}$' }
            make: { type: string, minLength: 1, maxLength: 60 }
            model: { type: string, minLength: 1, maxLength: 60 }
            year: { type: integer, minimum: 1950, maximum: 2100 }
            mileage: { type: integer, minimum: 0 }
            mileageUnit: { type: string, enum: [KM, MI] }
            status: { type: string, enum: [ACTIVE, ARCHIVED] }
            lastCompletedJob: { $ref: '#/components/schemas/ServiceHistoryItem' }
            nextService: { $ref: '#/components/schemas/NextService' }
    VehicleCreateRequest:
      type: object
      required: [customerId, plate, vin, make, model, year, mileage, mileageUnit]
      additionalProperties: false
      properties: { customerId: { $ref: '#/components/schemas/Uuid' }, plate: { type: string, minLength: 2, maxLength: 20 }, vin: { type: string, pattern: '^[A-HJ-NPR-Z0-9]{17}$' }, make: { type: string, minLength: 1, maxLength: 60 }, model: { type: string, minLength: 1, maxLength: 60 }, year: { type: integer, minimum: 1950, maximum: 2100 }, mileage: { type: integer, minimum: 0 }, mileageUnit: { type: string, enum: [KM, MI] } }
    VehicleUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { plate: { type: string, minLength: 2, maxLength: 20 }, mileage: { type: integer, minimum: 0 }, status: { type: string, enum: [ACTIVE, ARCHIVED] } }
    ServiceReminder:
      allOf:
        - $ref: '#/components/schemas/RecordMeta'
        - type: object
          required: [vehicleId, title, status]
          properties: { vehicleId: { $ref: '#/components/schemas/Uuid' }, title: { type: string, maxLength: 160 }, dueDate: { $ref: '#/components/schemas/DateOnly' }, dueMileage: { type: integer, minimum: 0 }, status: { type: string, enum: [OPEN, DONE, CANCELLED] }, completedAt: { $ref: '#/components/schemas/Timestamp' }, notes: { type: string, maxLength: 500 } }
    ServiceReminderCreateRequest:
      type: object
      required: [title]
      additionalProperties: false
      anyOf: [ { required: [dueDate] }, { required: [dueMileage] } ]
      properties: { title: { type: string, minLength: 1, maxLength: 160 }, dueDate: { $ref: '#/components/schemas/DateOnly' }, dueMileage: { type: integer, minimum: 0 }, notes: { type: string, maxLength: 500 } }
    ServiceReminderUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties: { title: { type: string, minLength: 1, maxLength: 160 }, dueDate: { $ref: '#/components/schemas/DateOnly' }, dueMileage: { type: integer, minimum: 0 }, status: { type: string, enum: [DONE, CANCELLED] }, notes: { type: string, maxLength: 500 } }
...
    JobNumber: { type: string, pattern: '^JC-[0-9]{4}-[0-9]{6}$', description: "Server-generated, unique, human-readable. Always rendered LTR by clients." }
...
    ServiceType: { type: string, enum: [MAINTENANCE, REPAIR, DIAGNOSTIC, INSPECTION, OTHER] }
...
    VehiclePage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/Vehicle' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    ServiceHistoryPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/ServiceHistoryItem' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    ServiceReminderPage: { type: object, required: [items, page], properties: { items: { type: array, items: { $ref: '#/components/schemas/ServiceReminder' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...